+++ previous-versions/assess/src/codatassess/sdk.py
@@ -7,7 +7,7 @@ from .reports import Reports
 from .sdkconfiguration import SDKConfiguration
 from codatassess import utils
 from codatassess.models import shared
-from typing import Dict
+from typing import Callable, Dict, Union
 
 class CodatAssess:
     r"""Assess API: Codat's financial insights API
@@ -17,17 +17,17 @@ class CodatAssess:
 
     [See our OpenAPI spec](https://github.com/codatio/oas)
     """
+    reports: Reports
+    r"""Enriched reports and analyses of financial data"""
     data_integrity: DataIntegrity
     r"""Match mutable accounting data with immutable banking data to increase confidence in financial data"""
     excel_reports: ExcelReports
     r"""Downloadable reports"""
-    reports: Reports
-    r"""Enriched reports and analyses of financial data"""
 
     sdk_configuration: SDKConfiguration
 
     def __init__(self,
-                 auth_header: str,
+                 auth_header: Union[str,Callable[[], str]],
                  server_idx: int = None,
                  server_url: str = None,
                  url_params: Dict[str, str] = None,
@@ -37,7 +37,7 @@ class CodatAssess:
         """Instantiates the SDK configuring it with the provided parameters.
         
         :param auth_header: The auth_header required for authentication
-        :type auth_header: str
+        :type auth_header: Union[str,Callable[[], str]]
         :param server_idx: The index of the server to use for all operations
         :type server_idx: int
         :param server_url: The server URL to use for all operations
@@ -52,20 +52,18 @@ class CodatAssess:
         if client is None:
             client = requests_http.Session()
         
-        
-        security_client = utils.configure_security_client(client, shared.Security(auth_header = auth_header))
-        
+        security = shared.Security(auth_header = auth_header)
         
         if server_url is not None:
             if url_params is not None:
                 server_url = utils.template_url(server_url, url_params)
 
-        self.sdk_configuration = SDKConfiguration(client, security_client, server_url, server_idx, retry_config=retry_config)
+        self.sdk_configuration = SDKConfiguration(client, security, server_url, server_idx, retry_config=retry_config)
        
         self._init_sdks()
     
     def _init_sdks(self):
+        self.reports = Reports(self.sdk_configuration)
         self.data_integrity = DataIntegrity(self.sdk_configuration)
         self.excel_reports = ExcelReports(self.sdk_configuration)
-        self.reports = Reports(self.sdk_configuration)